# backend/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handlers.ProcessRequestInput:
    properties:
      status:
        description: Status de la demande (approved/rejected)
        enum:
        - approved
        - rejected
        example: approved
        type: string
    required:
    - status
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: Message d'erreur
        type: string
    type: object
...
  models.Merchant:
    properties:
      business_name:
        description: Nom de l'entreprise
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email_pro:
        description: Email valide requis
        type: string
      id:
        type: integer
      phone_number:
        description: Numéro de téléphone (optionnel, max 15 caractères)
        type: string
      siret:
        description: Numéro SIRET
        type: string
      updatedAt:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relation vers User (clé étrangère avec cascade)
      user_id:
        description: Relation 1 à 1 vers User
        type: integer
    required:
    - business_name
    - email_pro
    - siret
    type: object
  models.MerchantRequest:
    properties:
      business_name:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email_pro:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      siret:
        type: string
      status:
        description: pending, approved, rejected
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        description: Relation avec l'utilisateur qui fait la demande
        type: integer
    required:
    - business_name
    - email_pro
    - siret
    type: object
  models.Response:
    properties:
      data: {}
      message:
        example: Opération réussie
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: Validation d'email
        type: string
      full_name:
        description: Nom complet requis
        type: string
      id:
        type: integer
      is_admin:
        description: Est-ce un administrateur ?
        type: boolean
      password_hash:
        description: Hash du mot de passe
        type: string
      phone:
        description: Téléphone requis
        type: string
      updatedAt:
        type: string
    required:
    - email
    - full_name
    - password_hash
    - phone
    type: object
...
  requests.CreateMerchantRequest:
    properties:
      business_name:
        example: petit bateau
        type: string
      email_pro:
        example: merchant@example.com
        type: string
      phone_number:
        example: "+32452101010"
        type: string
      siret:
        example: "78467169500087"
        type: string
    required:
    - business_name
    - email_pro
    - siret
    type: object
...
  requests.UpdateMerchantRequest:
    properties:
      business_name:
        description: Nom de l'entreprise
        example: petit bateau update
        type: string
      email_pro:
        description: Email valide requis
        example: merchantupdate@example.com
        type: string
      phone_number:
        example: "+32452101010"
        type: string
      siret:
        description: Numéro SIRET
        example: "78467169500089"
        type: string
    required:
    - business_name
    - email_pro
    - siret
    type: object
...
info:
  contact: {}
paths:
  /api/admin/merchant-requests:
    get:
      description: Récupère toutes les demandes de marchand en attente (admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Liste des demandes en attente
          schema:
            items:
              $ref: '#/definitions/models.MerchantRequest'
            type: array
        "401":
          description: Non authentifié
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Non autorisé
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Erreur serveur
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer les demandes en attente
      tags:
      - Admin
  /api/admin/merchant-requests/{id}:
    put:
      consumes:
      - application/json
      description: Permet à un administrateur d'approuver ou rejeter une demande de
        marchand
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de la demande
        in: path
        name: id
        required: true
        type: integer
      - description: Statut de la demande (approved/rejected)
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/handlers.ProcessRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: Demande traitée avec succès
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: ID invalide ou statut invalide
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Non authentifié
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Non autorisé
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Erreur serveur
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Traiter une demande de marchand
      tags:
      - Admin
  /api/admin/merchants:
    get:
      description: Récupère tout les marchands (admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Liste des marchands
          schema:
            items:
              $ref: '#/definitions/models.Merchant'
            type: array
        "401":
          description: Non authentifié
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Non autorisé
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Erreur serveur
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer les marchands
      tags:
      - Admin
...
  /api/merchants:
    delete:
      consumes:
      - application/json
      description: Permet à un marchand de supprimer son compte
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Suppression d'un marchand
      tags:
      - Merchants
    get:
      description: Récupère les informations du marchand actuel
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Informations du marchand
          schema:
            $ref: '#/definitions/models.Merchant'
        "401":
          description: Non authentifié
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Non autorisé
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Marchand non trouvé
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Erreur serveur
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer information d'un marchand
      tags:
      - Merchants
    post:
      consumes:
      - application/json
      description: Permet à un utilisateur de soumettre une demande pour devenir marchand
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Données de la demande
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CreateMerchantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Créer une demande de marchand
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Permet à un marchand de mettre à jour ses informations
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Données du marchand
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update un marchand
      tags:
      - Merchants
...
swagger: "2.0"